Workaround terminal focus
There seems to be a bug in Ubuntu 12.04 preventing terminal windows from
noticing when they recieve focus. The cursor in the affected terminals
will look the same as in unfocus windows which is really annoying. I
changed from block cursors in the terminals and increased the brightness
of the focused window frame.  I will try out later releases on my home
computer to see if it has been solved, meanwhile this workaround should
be enough for work...