Updated to Awesome 3.4.15
I just updated my work computer to Ubuntu 14.04 (from 12.04).
As usual, various things broke and it took me half a day of debugging
before I could log in to some kind of graphical environment again.
However, I had to disable a lot of autostart applications.

After some further work, I've now managed to get all the auto-start
applications working again.  However, in order to clean things up a bit
I've moved most autostarts to ~/.xinitrc.  The nice thing with this is
that it practically eliminates the need for a run_once function, you can
restart awesome as much as you like but ~/.xinitrc is only run once.

Furthermore I aim to move the screen specific autostarts over to xinitrc
as well, but we'll see when that happens...