Added inkscape command to menu
There is a known bug when running Inkscape in Ubuntu 14.04:
https://bugs.launchpad.net/ubuntu/+source/inkscape/+bug/1385009		(+others)
There is a patch out but only for 15.04 or later.
For some reason, it seems to work when running Inkscape in a basic X-session.
This is (briefly) what my script does:
* Start xephyr as screen :2
* launch xfwm4 on :2.0
* launch inkscape on :2.0
The script is run from the main menu under applications.